requirements.txt
Django==1.8.5
Whoosh==2.7.0
argparse==1.2.1
django-haystack==2.4.0
requests==2.8.1
wsgiref==0.1.2
